requirements.txt
openpyxl==3.1.2
python-docx==1.1.0
google-generativeai>=0.8.0
python-dotenv==1.0.0